pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "webpage analyzer application"
authors = ["Dmitry Gorokhov <gorokhov86@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
repository = "https://github.com/DmGorokhov/python-project-83.git"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
install = "^1.3.5"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
